+++ tracker/src/plausible.js
@@ -1,33 +1,16 @@
+import { init as initEngagementTracking, prePageviewTrigger, postPageviewTrigger, onPageviewIgnored } from './engagement'
+import { sendRequest } from './networking'
+import { init as initConfig, config, scriptEl } from './config'
+
 var location = window.location
 var document = window.document
-
-if (COMPILE_COMPAT) {
-  var scriptEl = document.getElementById('plausible');
-} else {
-  var scriptEl = document.currentScript;
-}
-
-var config = {}
-
-if (COMPILE_CONFIG) {
-  config = "<%= @config_js %>"
-}
-
-var endpoint
-var dataDomain
-var autoCapturePageviews = !COMPILE_MANUAL
 
 // Exported public function
 function trigger(eventName, options) {
   var isPageview = eventName === 'pageview'
 
-  if (isPageview && listeningOnEngagement) {
-    // If we're listening on engagement already, at least one pageview
-    // has been sent by the current script (i.e. it's most likely a SPA).
-    // Trigger an engagement marking the "exit from the previous page".
-    triggerEngagement()
-    currentDocumentHeight = getDocumentHeight()
-    maxScrollDepthPx = getCurrentScrollDepthPx()
+  if (isPageview) {
+    prePageviewTrigger()
   }
 
   if (!(COMPILE_LOCAL && (!COMPILE_CONFIG || config.captureOnLocalhost))) {
@@ -72,7 +55,7 @@ function trigger(eventName, options) {
   payload.n = eventName
   payload.v = COMPILE_TRACKER_SCRIPT_VERSION
 
-  if (COMPILE_MANUAL || !autoCapturePageviews) {
+  if (!COMPILE_MANUAL || (COMPILE_CONFIG && config.autoCapturePageviews)) {
     var customURL = options && options.u
 
     payload.u = customURL ? customURL : location.href
@@ -80,7 +63,7 @@ function trigger(eventName, options) {
     payload.u = location.href
   }
 
-  payload.d = dataDomain
+  payload.d = config.domain
   payload.r = document.referrer || null
   if (options && options.meta) {
     payload.m = JSON.stringify(options.meta)
@@ -130,178 +113,22 @@ function trigger(eventName, options) {
   }
 
   if (isPageview) {
-    currentEngagementIgnored = false
-    currentEngagementURL = payload.u
-    currentEngagementProps = payload.p
-    currentEngagementMaxScrollDepth = -1
-    currentEngagementTime = 0
-    runningEngagementStart = Date.now()
-    registerEngagementListener()
+    postPageviewTrigger(payload)
   }
 
-  sendRequest(endpoint, payload, options)
+  sendRequest(config.endpoint, payload, options)
 }
 
-function sendRequest(endpoint, payload, options) {
-  if (COMPILE_COMPAT) {
-    var request = new XMLHttpRequest();
-    request.open('POST', endpoint, true);
-    request.setRequestHeader('Content-Type', 'text/plain');
-
-    request.send(JSON.stringify(payload));
-
-    request.onreadystatechange = function () {
-      if (request.readyState === 4) {
-        options && options.callback && options.callback({ status: request.status })
-      }
-    }
-  } else {
-    if (window.fetch) {
-      fetch(endpoint, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'text/plain'
-        },
-        keepalive: true,
-        body: JSON.stringify(payload)
-      }).then(function (response) {
-        options && options.callback && options.callback({ status: response.status })
-      }).catch(function () { })
-    }
-  }
-}
-
-/* Engagement tracking variables and functions */
-
-// Multiple pageviews might be sent by the same script when the page
-// uses client-side routing (e.g. hash or history-based). This flag
-// prevents registering multiple listeners in those cases.
-var listeningOnEngagement = false
-
-var currentEngagementIgnored
-var currentEngagementURL = location.href
-var currentEngagementProps = {}
-var currentEngagementMaxScrollDepth = -1
-
-// Timestamp indicating when this particular page last became visible.
-// Reset during pageviews, set to 0 when page is closed.
-var runningEngagementStart = 0
-
-// When page is hidden, this 'engaged' time is saved to this variable
-var currentEngagementTime = 0
-
-function triggerEngagement() {
-  var engagementTime = getEngagementTime()
-
-  /*
-  We send engagements if there's new relevant engagement information to share:
-  - If the user has scrolled more than the previously sent max scroll depth.
-  - If the user has been engaged for more than 3 seconds since the last engagement event.
-
-  The first engagement event is always sent due to containing at least the initial scroll depth.
-
-  Also, we don't send engagements if the current pageview is ignored (onIgnoredEvent)
-  */
-  if (!currentEngagementIgnored && (currentEngagementMaxScrollDepth < maxScrollDepthPx || engagementTime >= 3000)) {
-    currentEngagementMaxScrollDepth = maxScrollDepthPx
-
-    var payload = {
-      n: 'engagement',
-      sd: Math.round((maxScrollDepthPx / currentDocumentHeight) * 100),
-      d: dataDomain,
-      u: currentEngagementURL,
-      p: currentEngagementProps,
-      e: engagementTime,
-      v: COMPILE_TRACKER_SCRIPT_VERSION
-    }
-
-    // Reset current engagement time metrics. They will restart upon when page becomes visible or the next SPA pageview
-    runningEngagementStart = 0
-    currentEngagementTime = 0
-
-    if (COMPILE_HASH && (!COMPILE_CONFIG || config.hashBasedRouting)) {
-      payload.h = 1
-    }
-
-    sendRequest(endpoint, payload)
-  }
-}
-
-function onVisibilityChange() {
-  if (document.visibilityState === 'visible' && document.hasFocus() && runningEngagementStart === 0) {
-    runningEngagementStart = Date.now()
-  } else if (document.visibilityState === 'hidden' || !document.hasFocus()) {
-    // Tab went back to background or lost focus. Save the engaged time so far
-    currentEngagementTime = getEngagementTime()
-    runningEngagementStart = 0
-
-    triggerEngagement()
-  }
-}
-
-function registerEngagementListener() {
-  if (!listeningOnEngagement) {
-    // Only register visibilitychange listener only after initial page load and pageview
-    document.addEventListener('visibilitychange', onVisibilityChange)
-    window.addEventListener('blur', onVisibilityChange)
-    window.addEventListener('focus', onVisibilityChange)
-    listeningOnEngagement = true
-  }
-}
-
-function getEngagementTime() {
-  if (runningEngagementStart) {
-    return currentEngagementTime + (Date.now() - runningEngagementStart)
-  } else {
-    return currentEngagementTime
-  }
-}
-
-
-var currentDocumentHeight
-var maxScrollDepthPx
-
-function getDocumentHeight() {
-  var body = document.body || {}
-  var el = document.documentElement || {}
-  return Math.max(
-    body.scrollHeight || 0,
-    body.offsetHeight || 0,
-    body.clientHeight || 0,
-    el.scrollHeight || 0,
-    el.offsetHeight || 0,
-    el.clientHeight || 0
-  )
-}
-
-function getCurrentScrollDepthPx() {
-  var body = document.body || {}
-  var el = document.documentElement || {}
-  var viewportHeight = window.innerHeight || el.clientHeight || 0
-  var scrollTop = window.scrollY || el.scrollTop || body.scrollTop || 0
-
-  return currentDocumentHeight <= viewportHeight ? currentDocumentHeight : scrollTop + viewportHeight
-}
 
 function onIgnoredEvent(eventName, options, reason) {
   if (reason) console.warn('Ignoring Event: ' + reason);
   options && options.callback && options.callback()
 
   if (eventName === 'pageview') {
-    currentEngagementIgnored = true
+    onPageviewIgnored()
   }
 }
 
-function defaultEndpoint() {
-  if (COMPILE_COMPAT) {
-    var pathArray = scriptEl.src.split('/');
-    var protocol = pathArray[0];
-    var host = pathArray[2];
-    return protocol + '//' + host + '/api/event';
-  } else {
-    return new URL(scriptEl.src).origin + '/api/event'
-  }
-}
 
 function init(overrides) {
   if (COMPILE_CONFIG && window.plausible && window.plausible.l) {
@@ -309,42 +136,10 @@ function init(overrides) {
     return
   }
 
-  // Explicitly set dataDomain before any overrides are applied as `plausible-web` does not support overriding it
-  dataDomain = COMPILE_CONFIG ? config.domain : scriptEl.getAttribute('data-domain')
+  initConfig(overrides)
+  initEngagementTracking()
 
-  if (COMPILE_CONFIG) {
-    Object.assign(config, overrides)
-    autoCapturePageviews = config.autoCapturePageviews !== false
-  }
-
-  endpoint = COMPILE_CONFIG ? config.endpoint : (scriptEl.getAttribute('data-api') || defaultEndpoint())
-
-  currentDocumentHeight = getDocumentHeight()
-  maxScrollDepthPx = getCurrentScrollDepthPx()
-
-  window.addEventListener('load', function () {
-    currentDocumentHeight = getDocumentHeight()
-
-    // Update the document height again after every 200ms during the
-    // next 3 seconds. This makes sure dynamically loaded content is
-    // also accounted for.
-    var count = 0
-    var interval = setInterval(function () {
-      currentDocumentHeight = getDocumentHeight()
-      if (++count === 15) { clearInterval(interval) }
-    }, 200)
-  })
-
-  document.addEventListener('scroll', function () {
-    currentDocumentHeight = getDocumentHeight()
-    var currentScrollDepthPx = getCurrentScrollDepthPx()
-
-    if (currentScrollDepthPx > maxScrollDepthPx) {
-      maxScrollDepthPx = currentScrollDepthPx
-    }
-  })
-
-  if (!COMPILE_MANUAL || autoCapturePageviews) {
+  if (!COMPILE_MANUAL || (COMPILE_CONFIG && config.autoCapturePageviews)) {
     var lastPage;
 
     function page(isSPANavigation) {
